Refactor AbstractBuildScanPluginPerformanceTest
