+++ platforms/enterprise/enterprise-plugin-performance/src/testFixtures/groovy/org/gradle/performance/AbstractBuildScanPluginPerformanceTest.groovy
@@ -22,6 +22,7 @@ import org.gradle.integtests.fixtures.versions.ReleasedVersionDistributions
 import org.gradle.performance.fixture.BuildExperimentSpec
 import org.gradle.performance.fixture.BuildScanPerformanceTestRunner
 import org.gradle.performance.fixture.GradleBuildExperimentRunner
+import org.gradle.performance.fixture.GradleBuildExperimentSpec
 import org.gradle.performance.fixture.GradleInvocationSpec
 import org.gradle.performance.fixture.PerformanceTestIdProvider
 import org.gradle.performance.measure.Amount
@@ -90,8 +91,17 @@ class AbstractBuildScanPluginPerformanceTest extends AbstractPerformanceTest {
                 invocation.buildLog(new File(builder.workingDirectory, "build.log"))
                 invocation.distribution(distribution)
             }
+
+            @Override
+            protected void configureGradleSpec(GradleBuildExperimentSpec.GradleBuilder builder) {
+                super.configureGradleSpec(builder)
+                AbstractBuildScanPluginPerformanceTest.this.configureGradleSpec(builder)
+            }
         }
         performanceTestIdProvider.setTestSpec(runner)
+    }
+
+    protected void configureGradleSpec(GradleBuildExperimentSpec.GradleBuilder builder) {
     }
 
     private static resolvePluginVersion() {
